nginx/requirements.txt
grpcio
grpcio-health-checking==1.48.2